# openapi/minimovie-api.yaml
openapi: 3.0.3
info:
  title: MiniMovie API
  description: API for retrieving intuitive movie, TV series, and person information
  version: 1.0.0
servers:
  - url: https://api.minimovie.info
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /series/{seriesId}/seasons/{seasonNumber}:
    get:
      summary: Get season details
      description: Retrieve detailed information about a TV season including episodes and credits
      operationId: getSeason
      parameters:
        - name: seriesId
          in: path
          required: true
          description: TMDB series ID
          schema:
            type: integer
        - name: seasonNumber
          in: path
          required: true
          description: Season number
          schema:
            type: integer
      responses:
        "200":
          description: Season details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeasonDetails"
        "400":
          description: Invalid series ID or season number
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Season not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /users/me/progress/{seriesId}:
    get:
      summary: Get watch progress for a series
      description: Returns which episodes the user has watched for a given series
      operationId: getWatchProgress
      security:
        - sessionCookie: []
      parameters:
        - name: seriesId
          in: path
          required: true
          description: TMDB series ID
          schema:
            type: integer
      responses:
        "200":
          description: Watch progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchProgressResponse"
        "400":
          description: Invalid seriesId
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/me/progress:
    post:
      summary: Mark episode as watched
      description: >-
        Marks a specific episode as watched by creating a watch event.
        Also updates the associated watchlist item summary.
      operationId: markEpisodeWatched
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkEpisodeWatchedRequest"
      responses:
        "201":
          description: Episode marked as watched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchEvent"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Episode not found on TMDB
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Metadata resolution failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/me/progress/{id}:
    delete:
      summary: Unmark episode
      description: Removes a watch event, effectively unmarking the episode as watched
      operationId: unmarkEpisode
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          description: Watch event ID
          schema:
            type: string
      responses:
        "204":
          description: Episode unmarked
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Event not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/me/watch-events:
    get:
      summary: List watch events
      description: Returns the user's watch event history with optional filters
      operationId: listWatchEvents
      security:
        - sessionCookie: []
      parameters:
        - name: media_type
          in: query
          required: false
          schema:
            type: string
            enum:
              - movie
              - episode
              - series
              - season
        - name: media_id
          in: query
          required: false
          schema:
            type: integer
        - name: series_id
          in: query
          required: false
          schema:
            type: integer
        - name: dated_only
          in: query
          required: false
          description: Only return events with a watchedAt date
          schema:
            type: string
            enum:
              - "true"
        - name: limit
          in: query
          required: false
          description: Max number of events to return (default 100, max 500)
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
      responses:
        "200":
          description: Watch events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchEventListResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Create watch event
      description: >-
        Records a new watch event for a movie, series, or episode.
        Resolves metadata from TMDB and updates watchlist summary.
      operationId: createWatchEvent
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWatchEventRequest"
      responses:
        "201":
          description: Watch event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchEvent"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Media not found on TMDB
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Metadata resolution failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Guest API key used for public catalog routes and token exchange
    sessionCookie:
      type: apiKey
      in: cookie
      name: mm_session
      description: Session cookie set after OAuth login (production name is __Host-mm_session)
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: HTTP status text
        message:
          type: string
          description: Detailed error message
...
    SeasonDetails:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        overview:
          type: string
        posterPath:
          type: string
        seasonNumber:
          type: integer
        airDate:
          type: string
          format: date
        voteAverage:
          type: number
          format: float
        episodes:
          type: array
          items:
            $ref: "#/components/schemas/EpisodeSummary"
        whereToWatch:
          $ref: "#/components/schemas/WhereToWatch"
        credits:
          $ref: "#/components/schemas/Credits"
    EpisodeSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        overview:
          type: string
        episodeNumber:
          type: integer
        seasonNumber:
          type: integer
        airDate:
          type: string
          format: date
        runtime:
          type: integer
          description: Runtime in minutes
        stillPath:
          type: string
        voteAverage:
          type: number
          format: float
...
    Person:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        photoPath:
          type: string
        role:
          type: string
          description: Character name or job title
        order:
          type: integer
          description: Billing order (for cast)
        episodeCount:
          type: integer
          description: Number of episodes the person appeared in (for series/season aggregate credits)
        birthday:
          type: string
          format: date
        deathday:
          type: string
          format: date
        currentAge:
          type: integer
          description: Current age of the person (if alive)
        ageAtRelease:
          type: integer
          description: Age of the person at the time of release
        ageRange:
          type: string
          description: Age range during production (for series spanning multiple years)
...
    WhereToWatch:
      type: object
      properties:
        stream:
          type: array
          items:
            $ref: "#/components/schemas/WatchProvider"
        rent:
          type: array
          items:
            $ref: "#/components/schemas/WatchProvider"
        buy:
          type: array
          items:
            $ref: "#/components/schemas/WatchProvider"
        free:
          type: array
          items:
            $ref: "#/components/schemas/WatchProvider"
        ads:
          type: array
          items:
            $ref: "#/components/schemas/WatchProvider"
    WatchProvider:
      type: object
      properties:
        name:
          type: string
        logoPath:
          type: string
    Credits:
      type: object
      properties:
        cast:
          type: array
          items:
            $ref: "#/components/schemas/Person"
        directors:
          type: array
          items:
            $ref: "#/components/schemas/Person"
        writers:
          type: array
          items:
            $ref: "#/components/schemas/Person"
        producers:
          type: array
          items:
            $ref: "#/components/schemas/Person"
        composers:
          type: array
          items:
            $ref: "#/components/schemas/Person"
        cinematographers:
          type: array
          items:
            $ref: "#/components/schemas/Person"
        editors:
          type: array
          items:
            $ref: "#/components/schemas/Person"
        productionDesign:
          type: array
          items:
            $ref: "#/components/schemas/Person"
        costumeDesign:
          type: array
          items:
            $ref: "#/components/schemas/Person"
        casting:
          type: array
          items:
            $ref: "#/components/schemas/Person"
...
    WatchEvent:
      type: object
      properties:
        id:
          type: string
        mediaType:
          type: string
          enum:
            - movie
            - episode
            - series
            - season
        mediaId:
          type: integer
        mediaTitle:
          type: string
        seriesId:
          type: integer
          nullable: true
        seriesTitle:
          type: string
          nullable: true
        seasonNumber:
          type: integer
          nullable: true
        episodeNumber:
          type: integer
          nullable: true
        episodeCount:
          type: integer
          nullable: true
          description: Total episodes in the series/season (for series/season-level events)
        seasonCount:
          type: integer
          nullable: true
          description: Total seasons in the series (for series-level events)
        watchedAt:
          type: string
          format: date-time
          nullable: true
        timezone:
          type: string
        rewatchNumber:
          type: integer
        runtimeMinutes:
          type: integer
          nullable: true
        genres:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
    WatchEventListResponse:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/WatchEvent"
        total:
          type: integer
    CreateWatchEventRequest:
      type: object
      required:
        - mediaType
        - mediaId
      properties:
        mediaType:
          type: string
          enum:
            - movie
            - episode
            - series
            - season
        mediaId:
          type: integer
          minimum: 1
        seriesId:
          type: integer
          description: Required when mediaType is episode or season
        seasonNumber:
          type: integer
          description: Required when mediaType is episode or season
        episodeNumber:
          type: integer
          description: Required when mediaType is episode
        justWatched:
          type: boolean
          description: If true, sets watchedAt to the current time
          default: false
        timezone:
          type: string
          description: IANA timezone (defaults to UTC)
          default: UTC
    MarkEpisodeWatchedRequest:
      type: object
      required:
        - seriesId
        - seasonNumber
        - episodeNumber
      properties:
        seriesId:
          type: integer
        seasonNumber:
          type: integer
        episodeNumber:
          type: integer
        timezone:
          type: string
          description: IANA timezone (defaults to UTC)
          default: UTC
    WatchProgressResponse:
      type: object
      properties:
        seriesId:
          type: integer
        episodes:
          type: array
          items:
            $ref: "#/components/schemas/EpisodeProgress"
        total:
          type: integer
    EpisodeProgress:
      type: object
      properties:
        id:
          type: string
        seasonNumber:
          type: integer
          nullable: true
        episodeNumber:
          type: integer
          nullable: true
        watchedAt:
          type: string
          format: date-time
          nullable: true
...